feat: move Don't Sleep toggle back to activity rail above settings
Relocates the Don't Sleep button from the status bar into the activity
rail, restored to its original position above the Settings button. The
rail picks up the existing persistence path — the preference is saved
in general settings and re-applied on launch — and continues to sync
with the tray menu via the kkterm://dont-sleep-changed event.
